move redirect tests into redirect.test.tsx
