+++ packages/react-router/tests/loaders.test.tsx
@@ -114,132 +114,4 @@ describe('loaders are being called', () => {
     expect(nestedLoaderMock).toHaveBeenCalled()
     expect(nestedFooLoaderMock).toHaveBeenCalled()
   })
-
-  test('both are called on /nested/foo when redirected in "beforeLoad" from /about', async () => {
-    const nestedLoaderMock = vi.fn()
-    const nestedFooLoaderMock = vi.fn()
-
-    const rootRoute = createRootRoute({})
-    const indexRoute = createRoute({
-      getParentRoute: () => rootRoute,
-      path: '/',
-      component: () => {
-        return (
-          <div>
-            <h1>Index page</h1>
-            <Link to="/about">link to about</Link>
-          </div>
-        )
-      },
-    })
-    const aboutRoute = createRoute({
-      getParentRoute: () => rootRoute,
-      path: '/about',
-      beforeLoad: async () => {
-        await sleep(WAIT_TIME)
-        throw redirect({ to: '/nested/foo' })
-      },
-    })
-    const nestedRoute = createRoute({
-      getParentRoute: () => rootRoute,
-      path: '/nested',
-      loader: async () => {
-        await sleep(WAIT_TIME)
-        nestedLoaderMock('nested')
-      },
-    })
-    const fooRoute = createRoute({
-      getParentRoute: () => nestedRoute,
-      path: '/foo',
-      loader: async () => {
-        await sleep(WAIT_TIME)
-        nestedFooLoaderMock('foo')
-      },
-      component: () => <div>Nested Foo page</div>,
-    })
-    const routeTree = rootRoute.addChildren([
-      nestedRoute.addChildren([fooRoute]),
-      aboutRoute,
-      indexRoute,
-    ])
-    const router = await act(() => createRouter({ routeTree }))
-
-    await act(() => render(<RouterProvider router={router} />))
-
-    const linkToAbout = await screen.findByText('link to about')
-    await act(() => fireEvent.click(linkToAbout))
-
-    const fooElement = await screen.findByText('Nested Foo page')
-    expect(fooElement).toBeInTheDocument()
-
-    expect(router.state.location.href).toBe('/nested/foo')
-    expect(window.location.pathname).toBe('/nested/foo')
-
-    expect(nestedLoaderMock).toHaveBeenCalled()
-    expect(nestedFooLoaderMock).toHaveBeenCalled()
-  })
-
-  test('both are called on /nested/foo when redirected in "loader" from /about', async () => {
-    const nestedLoaderMock = vi.fn()
-    const nestedFooLoaderMock = vi.fn()
-
-    const rootRoute = createRootRoute({})
-    const indexRoute = createRoute({
-      getParentRoute: () => rootRoute,
-      path: '/',
-      component: () => {
-        return (
-          <div>
-            <h1>Index page</h1>
-            <Link to="/about">link to about</Link>
-          </div>
-        )
-      },
-    })
-    const aboutRoute = createRoute({
-      getParentRoute: () => rootRoute,
-      path: '/about',
-      loader: async () => {
-        await sleep(WAIT_TIME)
-        throw redirect({ to: '/nested/foo' })
-      },
-    })
-    const nestedRoute = createRoute({
-      getParentRoute: () => rootRoute,
-      path: '/nested',
-      loader: async () => {
-        await sleep(WAIT_TIME)
-        nestedLoaderMock('nested')
-      },
-    })
-    const fooRoute = createRoute({
-      getParentRoute: () => nestedRoute,
-      path: '/foo',
-      loader: async () => {
-        await sleep(WAIT_TIME)
-        nestedFooLoaderMock('foo')
-      },
-      component: () => <div>Nested Foo page</div>,
-    })
-    const routeTree = rootRoute.addChildren([
-      nestedRoute.addChildren([fooRoute]),
-      aboutRoute,
-      indexRoute,
-    ])
-    const router = await act(() => createRouter({ routeTree }))
-
-    await act(() => render(<RouterProvider router={router} />))
-
-    const linkToAbout = await screen.findByText('link to about')
-    await act(() => fireEvent.click(linkToAbout))
-
-    const fooElement = await screen.findByText('Nested Foo page')
-    expect(fooElement).toBeInTheDocument()
-
-    expect(router.state.location.href).toBe('/nested/foo')
-    expect(window.location.pathname).toBe('/nested/foo')
-
-    expect(nestedLoaderMock).toHaveBeenCalled()
-    expect(nestedFooLoaderMock).toHaveBeenCalled()
-  })
 })
